BFD: Fix invalid read from pollfd array #2
Previous fix (5424fee) isn't correct, an array pfd[] is allocated well
to number of (loop->sock_num + 1) items.

The problem was that, there is possible to be called sockets_add()
between sockets_prepare() and sockets_fire() using poll loop in
birdloop_main(), so we need to use as last item in pfd[] the variable
`used' from buffer structure.